Moved Rename Recovery test cases to seperate class
