+++ hadoop-tools/hadoop-azure/src/test/java/org/apache/hadoop/fs/azurebfs/ITestAzureBlobFileSystemRename.java
@@ -24,10 +24,8 @@ import java.net.URI;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.AccessDeniedException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
@@ -51,7 +49,6 @@ import org.apache.hadoop.fs.azurebfs.constants.HttpHeaderConfigurations;
 import org.apache.hadoop.fs.azurebfs.contracts.exceptions.AbfsDriverException;
 import org.apache.hadoop.fs.azurebfs.contracts.exceptions.AbfsRestOperationException;
 import org.apache.hadoop.fs.azurebfs.contracts.exceptions.AzureBlobFileSystemException;
-import org.apache.hadoop.fs.azurebfs.contracts.services.AzureServiceErrorCode;
 import org.apache.hadoop.fs.azurebfs.security.ContextEncryptionAdapter;
 import org.apache.hadoop.fs.azurebfs.services.AbfsBlobClient;
 import org.apache.hadoop.fs.azurebfs.services.AbfsClient;
@@ -94,13 +91,10 @@ import static org.apache.hadoop.fs.azurebfs.constants.ConfigurationKeys.FS_AZURE
 import static org.apache.hadoop.fs.azurebfs.constants.ConfigurationKeys.FS_AZURE_ENABLE_CLIENT_TRANSACTION_ID;
 import static org.apache.hadoop.fs.azurebfs.constants.ConfigurationKeys.FS_AZURE_LEASE_THREADS;
 import static org.apache.hadoop.fs.azurebfs.constants.ConfigurationKeys.FS_AZURE_PRODUCER_QUEUE_MAX_SIZE;
-import static org.apache.hadoop.fs.azurebfs.contracts.services.AzureServiceErrorCode.BLOB_ALREADY_EXISTS;
-import static org.apache.hadoop.fs.azurebfs.contracts.services.AzureServiceErrorCode.BLOB_PATH_NOT_FOUND;
 import static org.apache.hadoop.fs.azurebfs.constants.HttpHeaderConfigurations.X_MS_CLIENT_TRANSACTION_ID;
 import static org.apache.hadoop.fs.azurebfs.constants.HttpHeaderConfigurations.X_MS_RESOURCE_TYPE;
 import static org.apache.hadoop.fs.azurebfs.contracts.services.AzureServiceErrorCode.COPY_BLOB_ABORTED;
 import static org.apache.hadoop.fs.azurebfs.contracts.services.AzureServiceErrorCode.COPY_BLOB_FAILED;
-import static org.apache.hadoop.fs.azurebfs.contracts.services.AzureServiceErrorCode.PATH_NOT_FOUND;
 import static org.apache.hadoop.fs.azurebfs.contracts.services.AzureServiceErrorCode.SOURCE_PATH_NOT_FOUND;
 import static org.apache.hadoop.fs.azurebfs.services.AbfsClientTestUtil.mockAddClientTransactionIdToHeader;
 import static org.apache.hadoop.fs.azurebfs.services.AbfsErrors.ERR_RENAME_RECOVERY;
@@ -125,10 +119,6 @@ public class ITestAzureBlobFileSystemRename extends
   private static final int BLOB_COUNT = 11;
 
   private static final int TOTAL_FILES = 25;
-
-  private static final int TOTAL_THREADS_IN_POOL = 5;
-
-  private static final int FAILED_CALL = 15;
 
   public ITestAzureBlobFileSystemRename() throws Exception {
     super();
@@ -1748,81 +1738,14 @@ public class ITestAzureBlobFileSystemRename extends
   private void renameDir(AzureBlobFileSystem currentFs, String producerQueueSize,
       String consumerMaxLag, String maxThread, Path src, Path dst)
       throws Exception {
-    Configuration config = createConfig(producerQueueSize, consumerMaxLag, maxThread);
-    try (AzureBlobFileSystem fs = (AzureBlobFileSystem) FileSystem.newInstance(currentFs.getUri(), config)) {
-      fs.rename(src, dst);
-      validateRename(fs, src, dst, false, true, false);
-    }
-  }
-
-  /**
-   * Helper method to create the configuration for the AzureBlobFileSystem.
-   *
-   * @param producerQueueSize Maximum size of the producer queue.
-   * @param consumerMaxLag Maximum lag allowed for the consumer.
-   * @param maxThread Maximum threads for the rename operation.
-   * @return The configuration object.
-   */
-  private Configuration createConfig(String producerQueueSize, String consumerMaxLag, String maxThread) {
     Configuration config = new Configuration(this.getRawConfiguration());
     config.set(FS_AZURE_PRODUCER_QUEUE_MAX_SIZE, producerQueueSize);
     config.set(FS_AZURE_CONSUMER_MAX_LAG, consumerMaxLag);
     config.set(FS_AZURE_BLOB_DIR_RENAME_MAX_THREAD, maxThread);
-    return config;
-  }
-
-  /**
-   * Helper method to validate that the rename was successful and that the destination exists.
-   *
-   * @param fs The AzureBlobFileSystem instance to check the existence on.
-   * @param dst The destination path.
-   * @param src The source path.
-   * @throws IOException If an I/O error occurs during the validation.
-   */
-  private void validateRename(AzureBlobFileSystem fs, Path src, Path dst,
-      boolean isSrcExist, boolean isDstExist, boolean isJsonExist) throws Exception {
-    // Validate pending JSON file status
-    assertPathStatus(fs,
-        new Path(src.getParent(), src.getName() + SUFFIX), isJsonExist,
-        "Pending JSON file");
-
-    // Validate source directory status
-    assertPathStatus(fs, src, isSrcExist, "Source directory");
-
-    // Validate destination directory status
-    assertPathStatus(fs, dst, isDstExist, "Destination directory");
-  }
-
-  /**
-   * Helper method to assert the status of a path in the AzureBlobFileSystem.
-   *
-   * @param fs The AzureBlobFileSystem instance to check the existence on.
-   * @param path The path to check.
-   * @param shouldExist Whether the path should exist or not.
-   * @param description A description for the assertion.
-   * @throws Exception If an error occurs during the assertion.
-   */
-  private void assertPathStatus(AzureBlobFileSystem fs, Path path,
-      boolean shouldExist, String description) throws Exception{
-    TracingContext tracingContext = getTestTracingContext(fs, true);
-    AbfsBlobClient client = ((AbfsBlobClient) fs.getAbfsClient());
-    if (shouldExist) {
-      int actualStatus = client.getPathStatus(
-              path.toUri().getPath(), tracingContext,
-              null, true)
-          .getResult().getStatusCode();
-      Assertions.assertThat(actualStatus)
-          .describedAs("%s should exists", description)
-          .isEqualTo(HTTP_OK);
-    } else {
-      AzureServiceErrorCode errorCode = LambdaTestUtils.intercept(
-          AbfsRestOperationException.class, () -> {
-        client.getPathStatus(path.toUri().getPath(), true,
-            tracingContext, null);
-      }).getErrorCode();
-      Assertions.assertThat(errorCode)
-          .describedAs("%s should not exists", description)
-          .isEqualTo(BLOB_PATH_NOT_FOUND);
+    try (AzureBlobFileSystem fs = (AzureBlobFileSystem) FileSystem.newInstance(currentFs.getUri(), config)) {
+      Assertions.assertThat(fs.rename(src, dst))
+          .describedAs("Rename should succeed.")
+          .isTrue();
     }
   }
 
@@ -1837,7 +1760,7 @@ public class ITestAzureBlobFileSystemRename extends
       Path dst = new Path("/hbase/A1/A3");
 
       // Create sample files in the source directory
-      createFiles(currentFs, src, TOTAL_FILES);
+      ITestAzureBlobFileSystemRenameRecovery.createFiles(currentFs, src, TOTAL_FILES);
 
       // Test renaming with different configurations
       renameDir(currentFs, "10", "5", "2", src, dst);
@@ -1853,133 +1776,6 @@ public class ITestAzureBlobFileSystemRename extends
           .contains(
               "Invalid configuration value detected for \"fs.azure.blob.dir.list.consumer.max.lag\". "
                   + "maxConsumptionLag should be lesser than maxSize");
-    }
-  }
-
-  /**
-   * Helper method to create files in the given directory.
-   *
-   * @param fs The AzureBlobFileSystem instance to use for file creation.
-   * @param src The source path (directory).
-   * @param numFiles The number of files to create.
-   * @throws ExecutionException, InterruptedException If an error occurs during file creation.
-   */
-  private void createFiles(AzureBlobFileSystem fs, Path src, int numFiles)
-      throws ExecutionException, InterruptedException {
-    ExecutorService executorService = Executors.newFixedThreadPool(TOTAL_THREADS_IN_POOL);
-    List<Future> futures = new ArrayList<>();
-    for (int i = 0; i < numFiles; i++) {
-      final int iter = i;
-      Future future = executorService.submit(() ->
-          fs.create(new Path(src, "file" + iter + ".txt")));
-      futures.add(future);
-    }
-    for (Future future : futures) {
-      future.get();
-    }
-    executorService.shutdown();
-  }
-
-  /**
-   * Tests renaming a directory with a failure during the copy operation.
-   * Simulates an error when copying on the 6th call.
-   */
-  @Test
-  public void testRenameCopyFailureInBetween() throws Exception {
-    try (AzureBlobFileSystem fs = Mockito.spy(this.getFileSystem(
-        createConfig("5", "3", "2")))) {
-      assumeBlobServiceType();
-      AbfsBlobClient client = (AbfsBlobClient) addSpyHooksOnClient(fs);
-      fs.getAbfsStore().setClient(client);
-      Path src = new Path("/hbase/A1/A2");
-      Path dst = new Path("/hbase/A1/A3");
-
-      // Create sample files in the source directory
-      createFiles(fs, src, TOTAL_FILES);
-
-      // Track the number of copy operations
-      AtomicInteger copyCall = new AtomicInteger(0);
-      renameCrashInBetween(client, copyCall);
-      Assertions.assertThat(fs.rename(src, dst))
-          .describedAs("Rename should crash in between.")
-          .isFalse();
-
-      // Validate copy operation count
-      Assertions.assertThat(copyCall.get())
-          .describedAs("Copy operation count should be less than 10.")
-          .isLessThan(TOTAL_FILES);
-
-      // Assertions to validate renamed destination and source
-      validateRename(fs, src, dst, true, true, true);
-
-      // Validate that rename redo operation was triggered
-      copyCall.set(0);
-      triggerRenameRecovery(fs, src);
-
-      Assertions.assertThat(copyCall.get())
-          .describedAs("Copy operation count should be greater than 0.")
-          .isGreaterThan(0);
-
-      // Validate final state of destination and source
-      validateRename(fs, src, dst, false, true, false);
-    }
-  }
-
-  /**
-   * Tests renaming a directory with a failure during the delete operation.
-   * Simulates an error on the 6th delete operation and verifies the behavior.
-   */
-  @Test
-  public void testRenameDeleteFailureInBetween() throws Exception {
-    try (AzureBlobFileSystem fs = Mockito.spy(this.getFileSystem(
-        createConfig("5", "3", "2")))) {
-      assumeBlobServiceType();
-      AbfsBlobClient client = (AbfsBlobClient) addSpyHooksOnClient(fs);
-      fs.getAbfsStore().setClient(client);
-      Path src = new Path("/hbase/A1/A2");
-      Path dst = new Path("/hbase/A1/A3");
-
-      // Create sample files in the source directory
-      createFiles(fs, src, TOTAL_FILES);
-
-      // Track the number of delete operations
-      AtomicInteger deleteCall = new AtomicInteger(0);
-      Mockito.doAnswer(deleteRequest -> {
-        if (deleteCall.get() == FAILED_CALL) {
-          throw new AbfsRestOperationException(
-              BLOB_PATH_NOT_FOUND.getStatusCode(),
-              BLOB_PATH_NOT_FOUND.getErrorCode(),
-              BLOB_PATH_NOT_FOUND.getErrorMessage(),
-              new Exception());
-        }
-        deleteCall.incrementAndGet();
-        return deleteRequest.callRealMethod();
-      }).when(client).deleteBlobPath(Mockito.any(Path.class),
-          Mockito.anyString(), Mockito.any(TracingContext.class));
-
-      Assertions.assertThat(fs.rename(src, dst))
-          .describedAs("Rename should crash in between.")
-          .isFalse();
-
-      // Validate delete operation count
-      Assertions.assertThat(deleteCall.get())
-          .describedAs("Delete operation count should be less than 10.")
-          .isLessThan(TOTAL_FILES);
-
-      // Assertions to validate renamed destination and source
-      validateRename(fs, src, dst, true, true, true);
-
-      // Validate that rename redo operation was triggered
-      deleteCall.set(0);
-      triggerRenameRecovery(fs, src);
-
-      Assertions.assertThat(deleteCall.get())
-          .describedAs("Delete operation count should be greater than 0.")
-          .isGreaterThan(0);
-
-      // Validate final state of destination and source
-      // Validate that delete redo operation was triggered
-      validateRename(fs, src, dst, false, true, false);
     }
   }
 
@@ -2101,325 +1897,6 @@ public class ITestAzureBlobFileSystemRename extends
   }
 
   /**
-   * Helper method to create a json file.
-   * @param path parent path
-   * @param renameJson rename json path
-   * @return file system
-   * @throws IOException in case of failure
-   */
-  public AzureBlobFileSystem createJsonFile(Path path, Path renameJson)
-      throws IOException {
-    final AzureBlobFileSystem fs = Mockito.spy(this.getFileSystem());
-    assumeBlobServiceType();
-    AzureBlobFileSystemStore store = Mockito.spy(fs.getAbfsStore());
-    Mockito.doReturn(store).when(fs).getAbfsStore();
-    AbfsClient client = Mockito.spy(store.getClient());
-    Mockito.doReturn(client).when(store).getClient();
-
-    fs.setWorkingDirectory(new Path(ROOT_PATH));
-    fs.create(new Path(path, "file.txt"));
-
-    VersionedFileStatus fileStatus
-        = (VersionedFileStatus) fs.getFileStatus(path);
-
-    new RenameAtomicity(path, new Path("/hbase/test4"),
-        renameJson, getTestTracingContext(fs, true),
-        fileStatus.getEtag(), client)
-        .preRename();
-
-    Assertions.assertThat(fs.exists(renameJson))
-        .describedAs("Rename Pending Json file should exist.")
-        .isTrue();
-
-    return fs;
-  }
-
-  /**
-   * Test case to verify crash recovery with a single child folder.
-   *
-   * This test simulates a scenario where a pending rename JSON file exists for a single child folder
-   * under the parent directory. It ensures that when listing the files in the parent directory,
-   * only the child folder (with the pending rename JSON file) is returned, and no additional files are listed.
-   *
-   * @throws Exception if any error occurs during the test execution
-   */
-  @Test
-  public void testListCrashRecoveryWithSingleChildFolder() throws Exception {
-    AzureBlobFileSystem fs = null;
-    try {
-      Path path = new Path("/hbase/A1/A2");
-      Path renameJson = new Path(path.getParent(), path.getName() + SUFFIX);
-      fs = createJsonFile(path, renameJson);
-
-      FileStatus[] fileStatuses = fs.listStatus(new Path("/hbase/A1"));
-
-      Assertions.assertThat(fileStatuses.length)
-          .describedAs("List should return 1 file")
-          .isEqualTo(1);
-    } finally {
-      if (fs != null) {
-        fs.close();
-      }
-    }
-  }
-
-  /**
-   * Test case to verify crash recovery with multiple child folders.
-   *
-   * This test simulates a scenario where a pending rename JSON file exists, and multiple files are
-   * created in the parent directory. It ensures that when listing the files in the parent directory,
-   * the correct number of files is returned, including the pending rename JSON file.
-   *
-   * @throws Exception if any error occurs during the test execution
-   */
-  @Test
-  public void testListCrashRecoveryWithMultipleChildFolder() throws Exception {
-    AzureBlobFileSystem fs = null;
-    try {
-      Path path = new Path("/hbase/A1/A2");
-      Path renameJson = new Path(path.getParent(), path.getName() + SUFFIX);
-      fs = createJsonFile(path, renameJson);
-
-      fs.create(new Path("/hbase/A1/file1.txt"));
-      fs.create(new Path("/hbase/A1/file2.txt"));
-
-      FileStatus[] fileStatuses = fs.listStatus(new Path("/hbase/A1"));
-
-      Assertions.assertThat(fileStatuses.length)
-          .describedAs("List should return 3 files")
-          .isEqualTo(3);
-    } finally {
-      if (fs != null) {
-        fs.close();
-      }
-    }
-  }
-
-  /**
-   * Test case to verify crash recovery with a pending rename JSON file.
-   *
-   * This test simulates a scenario where a pending rename JSON file exists in the parent directory,
-   * and it ensures that after the deletion of the target directory and creation of new files,
-   * the listing operation correctly returns the remaining files without considering the pending rename.
-   *
-   * @throws Exception if any error occurs during the test execution
-   */
-  @Test
-  public void testListCrashRecoveryWithPendingJsonFile() throws Exception {
-    AzureBlobFileSystem fs = null;
-    try {
-      Path path = new Path("/hbase/A1/A2");
-      Path renameJson = new Path(path.getParent(), path.getName() + SUFFIX);
-      fs = createJsonFile(path, renameJson);
-
-      fs.delete(path, true);
-      fs.create(new Path("/hbase/A1/file1.txt"));
-      fs.create(new Path("/hbase/A1/file2.txt"));
-
-      FileStatus[] fileStatuses = fs.listStatus(path.getParent());
-
-      Assertions.assertThat(fileStatuses.length)
-          .describedAs("List should return 2 files")
-          .isEqualTo(2);
-    } finally {
-      if (fs != null) {
-        fs.close();
-      }
-    }
-  }
-
-  /**
-   * Test case to verify crash recovery when no pending rename JSON file exists.
-   *
-   * This test simulates a scenario where there is no pending rename JSON file in the directory.
-   * It ensures that the listing operation correctly returns all files in the parent directory, including
-   * those created after the rename JSON file is deleted.
-   *
-   * @throws Exception if any error occurs during the test execution
-   */
-  @Test
-  public void testListCrashRecoveryWithoutAnyPendingJsonFile() throws Exception {
-    AzureBlobFileSystem fs = null;
-    try {
-      Path path = new Path("/hbase/A1/A2");
-      Path renameJson = new Path(path.getParent(), path.getName() + SUFFIX);
-      fs = createJsonFile(path, renameJson);
-
-      fs.delete(renameJson, true);
-      fs.create(new Path("/hbase/A1/file1.txt"));
-      fs.create(new Path("/hbase/A1/file2.txt"));
-
-      FileStatus[] fileStatuses = fs.listStatus(path.getParent());
-
-      Assertions.assertThat(fileStatuses.length)
-          .describedAs("List should return 3 files")
-          .isEqualTo(3);
-    } finally {
-      if (fs != null) {
-        fs.close();
-      }
-    }
-  }
-
-  /**
-   * Test case to verify crash recovery when a pending rename JSON directory exists.
-   *
-   * This test simulates a scenario where a pending rename JSON directory exists, ensuring that the
-   * listing operation correctly returns all files in the parent directory without triggering a redo
-   * rename operation. It also checks that the directory with the suffix "-RenamePending.json" exists.
-   *
-   * @throws Exception if any error occurs during the test execution
-   */
-  @Test
-  public void testListCrashRecoveryWithPendingJsonDir() throws Exception {
-    try (AzureBlobFileSystem fs = Mockito.spy(this.getFileSystem())) {
-      assumeBlobServiceType();
-      AbfsBlobClient client = (AbfsBlobClient) addSpyHooksOnClient(fs);
-
-      Path path = new Path("/hbase/A1/A2");
-      Path renameJson = new Path(path.getParent(), path.getName() + SUFFIX);
-      fs.mkdirs(renameJson);
-
-      fs.create(new Path(path.getParent(), "file1.txt"));
-      fs.create(new Path(path, "file2.txt"));
-
-      AtomicInteger redoRenameCall = new AtomicInteger(0);
-      Mockito.doAnswer(answer -> {
-        redoRenameCall.incrementAndGet();
-        return answer.callRealMethod();
-      }).when(client).getRedoRenameAtomicity(Mockito.any(Path.class),
-          Mockito.anyInt(), Mockito.any(TracingContext.class));
-
-      FileStatus[] fileStatuses = fs.listStatus(path.getParent());
-
-      Assertions.assertThat(fileStatuses.length)
-          .describedAs("List should return 3 files")
-          .isEqualTo(3);
-
-      Assertions.assertThat(redoRenameCall.get())
-          .describedAs("No redo rename call should be made")
-          .isEqualTo(0);
-
-      Assertions.assertThat(
-              Arrays.stream(fileStatuses)
-                  .anyMatch(status -> renameJson.toUri().getPath().equals(status.getPath().toUri().getPath())))
-          .describedAs("Directory with suffix -RenamePending.json should exist.")
-          .isTrue();
-    }
-  }
-
-  /**
-   * Test case to verify crash recovery during listing with multiple pending rename JSON files.
-   *
-   * This test simulates a scenario where multiple pending rename JSON files exist, ensuring that
-   * crash recovery properly handles the situation. It verifies that two redo rename calls are made
-   * and that the list operation returns the correct number of paths.
-   *
-   * @throws Exception if any error occurs during the test execution
-   */
-  @Test
-  public void testListCrashRecoveryWithMultipleJsonFile() throws Exception {
-    AzureBlobFileSystem fs = null;
-    try {
-      Path path = new Path("/hbase/A1/A2");
-
-      // 1st Json file
-      Path renameJson = new Path(path.getParent(), path.getName() + SUFFIX);
-      fs = createJsonFile(path, renameJson);
-      AbfsBlobClient client = (AbfsBlobClient) addSpyHooksOnClient(fs);
-
-      // 2nd Json file
-      Path path2 = new Path("/hbase/A1/A3");
-      fs.create(new Path(path2, "file3.txt"));
-
-      Path renameJson2 = new Path(path2.getParent(), path2.getName() + SUFFIX);
-      VersionedFileStatus fileStatus
-          = (VersionedFileStatus) fs.getFileStatus(path2);
-
-      new RenameAtomicity(path2, new Path("/hbase/test4"),
-          renameJson2, getTestTracingContext(fs, true),
-          fileStatus.getEtag(), client).preRename();
-
-      fs.create(new Path(path, "file2.txt"));
-
-      AtomicInteger redoRenameCall = new AtomicInteger(0);
-      Mockito.doAnswer(answer -> {
-        redoRenameCall.incrementAndGet();
-        return answer.callRealMethod();
-      }).when(client).getRedoRenameAtomicity(Mockito.any(Path.class),
-          Mockito.anyInt(), Mockito.any(TracingContext.class));
-
-      FileStatus[] fileStatuses = fs.listStatus(path.getParent());
-
-      Assertions.assertThat(fileStatuses.length)
-          .describedAs("List should return 2 paths")
-          .isEqualTo(2);
-
-      Assertions.assertThat(redoRenameCall.get())
-          .describedAs("2 redo rename calls should be made")
-          .isEqualTo(2);
-    } finally {
-      if (fs != null) {
-        fs.close();
-      }
-    }
-  }
-
-  /**
-   * Test case to verify path status when a pending rename JSON file exists.
-   *
-   * This test simulates a scenario where a rename operation was pending, and ensures that
-   * the path status retrieval triggers a redo rename operation. The test also checks that
-   * the correct error code (`PATH_NOT_FOUND`) is returned.
-   *
-   * @throws Exception if any error occurs during the test execution
-   */
-  @Test
-  public void testGetPathStatusWithPendingJsonFile() throws Exception {
-    AzureBlobFileSystem fs = null;
-    try {
-      Path path = new Path("/hbase/A1/A2");
-      Path renameJson = new Path(path.getParent(), path.getName() + SUFFIX);
-      fs = createJsonFile(path, renameJson);
-
-      AbfsBlobClient client = (AbfsBlobClient) addSpyHooksOnClient(fs);
-
-      fs.create(new Path("/hbase/A1/file1.txt"));
-      fs.create(new Path("/hbase/A1/file2.txt"));
-
-      AbfsConfiguration conf = fs.getAbfsStore().getAbfsConfiguration();
-
-      AtomicInteger redoRenameCall = new AtomicInteger(0);
-      Mockito.doAnswer(answer -> {
-        redoRenameCall.incrementAndGet();
-        return answer.callRealMethod();
-      }).when(client).getRedoRenameAtomicity(Mockito.any(Path.class),
-          Mockito.anyInt(), Mockito.any(TracingContext.class));
-
-      TracingContext tracingContext = new TracingContext(
-          conf.getClientCorrelationId(), fs.getFileSystemId(),
-          FSOperationType.GET_FILESTATUS, TracingHeaderFormat.ALL_ID_FORMAT, null);
-
-      AzureServiceErrorCode azureServiceErrorCode = intercept(
-          AbfsRestOperationException.class, () -> client.getPathStatus(
-              path.toUri().getPath(), true,
-              tracingContext, null)).getErrorCode();
-
-      Assertions.assertThat(azureServiceErrorCode.getErrorCode())
-          .describedAs("Path had to be recovered from atomic rename operation.")
-          .isEqualTo(PATH_NOT_FOUND.getErrorCode());
-
-      Assertions.assertThat(redoRenameCall.get())
-          .describedAs("There should be one redo rename call")
-          .isEqualTo(1);
-    } finally {
-      if (fs != null) {
-        fs.close();
-      }
-    }
-  }
-
-  /**
    * Test case to verify path status when there is no pending rename JSON file.
    *
    * This test ensures that when no rename pending JSON file is present, the path status is
@@ -2526,71 +2003,6 @@ public class ITestAzureBlobFileSystemRename extends
       Assertions.assertThat(fs.exists(new Path(path.getParent(), path.getName() + SUFFIX)))
           .describedAs("Directory with suffix -RenamePending.json should exist.")
           .isTrue();
-    }
-  }
-
-  /**
-   * Test case to verify the behavior when the ETag of a file changes during a rename operation.
-   *
-   * This test simulates a scenario where the ETag of a file changes after the creation of a
-   * rename pending JSON file. The steps include:
-   * - Creating a rename pending JSON file with an old ETag.
-   * - Deleting the original directory for an ETag change.
-   * - Creating new files in the directory.
-   * - Verifying that the copy blob call is not triggered.
-   * - Verifying that the rename atomicity operation is called once.
-   *
-   * The test ensures that the system correctly handles the ETag change during the rename process.
-   *
-   * @throws Exception if any error occurs during the test execution
-   */
-  @Test
-  public void testETagChangedDuringRename() throws Exception {
-    AzureBlobFileSystem fs = null;
-    try {
-      assumeBlobServiceType();
-      Path path = new Path("/hbase/A1/A2");
-      Path renameJson = new Path(path.getParent(), path.getName() + SUFFIX);
-      // Create rename pending json file with old etag
-      fs = createJsonFile(path, renameJson);
-      AbfsBlobClient abfsBlobClient = (AbfsBlobClient) addSpyHooksOnClient(fs);
-      fs.getAbfsStore().setClient(abfsBlobClient);
-
-      // Delete the directory to change etag
-      fs.delete(path, true);
-
-      fs.create(new Path(path, "file1.txt"));
-      fs.create(new Path(path, "file2.txt"));
-      AtomicInteger numberOfCopyBlobCalls = new AtomicInteger(0);
-      Mockito.doAnswer(copyBlob -> {
-            numberOfCopyBlobCalls.incrementAndGet();
-            return copyBlob.callRealMethod();
-          })
-          .when(abfsBlobClient)
-          .copyBlob(Mockito.any(Path.class), Mockito.any(Path.class),
-              Mockito.nullable(String.class),
-              Mockito.any(TracingContext.class));
-
-      AtomicInteger numberOfRedoRenameAtomicityCalls = new AtomicInteger(0);
-      Mockito.doAnswer(redoRenameAtomicity -> {
-            numberOfRedoRenameAtomicityCalls.incrementAndGet();
-            return redoRenameAtomicity.callRealMethod();
-          })
-          .when(abfsBlobClient)
-          .getRedoRenameAtomicity(Mockito.any(Path.class), Mockito.anyInt(),
-              Mockito.any(TracingContext.class));
-      // Call list status to trigger rename redo
-      fs.listStatus(path.getParent());
-      Assertions.assertThat(numberOfRedoRenameAtomicityCalls.get())
-          .describedAs("There should be one call to getRedoRenameAtomicity")
-          .isEqualTo(1);
-      Assertions.assertThat(numberOfCopyBlobCalls.get())
-          .describedAs("There should be no copy blob call")
-          .isEqualTo(0);
-    } finally {
-      if (fs != null) {
-        fs.close();
-      }
     }
   }
 
@@ -2732,146 +2144,6 @@ public class ITestAzureBlobFileSystemRename extends
   }
 
   /**
-   * Tests renaming a directory with a failure during the copy operation.
-   * Since, destination path already exists, there will be adjustment in the
-   * destination path. After crash recovery, recovery should succeed even in the
-   * case when destination path already exists.
-   * Simulates an error when copying on the 6th call.
-   */
-  @Test
-  public void testRenameAndRecoveryWhenDestAlreadyExist() throws Exception {
-    try (AzureBlobFileSystem fs = Mockito.spy(this.getFileSystem(
-        createConfig("5", "3", "2")))) {
-      assumeBlobServiceType();
-      AbfsBlobClient client = (AbfsBlobClient) addSpyHooksOnClient(fs);
-      fs.getAbfsStore().setClient(client);
-      Path src = new Path("/hbase/A1/A2");
-      Path dst = new Path("/hbase/A1/A3");
-
-      // Create sample files in the source directory
-      createFiles(fs, src, TOTAL_FILES);
-      // Create the destination directory
-      fs.mkdirs(dst);
-
-      // Track the number of copy operations
-      AtomicInteger copyCall = new AtomicInteger(0);
-      renameCrashInBetween(client, copyCall);
-      Assertions.assertThat(fs.rename(src, dst))
-          .describedAs("Rename should crash in between.")
-          .isFalse();
-
-      // Validate copy operation count
-      Assertions.assertThat(copyCall.get())
-          .describedAs("Copy operation count should be less than 10.")
-          .isLessThan(TOTAL_FILES);
-
-      // Assertions to validate renamed destination and source
-      validateRename(fs, src, dst, true, true, true);
-
-      copyCall.set(0);
-      // List Status on src, this will internally do rename recovery
-      triggerRenameRecovery(fs, src);
-
-      Assertions.assertThat(copyCall.get())
-          .describedAs("Copy operation count should be greater than 0.")
-          .isGreaterThan(0);
-
-      // Validate final state of destination and source
-      validateRename(fs, src, dst, false, true, false);
-    }
-  }
-
-  /**
-   * Tests renaming a directory with a failure during the copy operation.
-   * Since, destination path already exists, there will be adjustment in the
-   * destination path. After crash recovery, recovery should succeed even in the
-   * case when destination path already exists.
-   * Simulates an error when copying on the 6th call.
-   */
-  @Test
-  public void testRenameRecoveryWithMarkerPresentInDest() throws Exception {
-    try (AzureBlobFileSystem fs = Mockito.spy(this.getFileSystem(
-        createConfig("5", "3", "2")))) {
-      assumeBlobServiceType();
-      AbfsBlobClient client = (AbfsBlobClient) addSpyHooksOnClient(fs);
-      fs.getAbfsStore().setClient(client);
-      Path src = new Path("/hbase/A1/A2");
-      Path dst = new Path("/hbase/A1/A3");
-
-      // Create sample files in the source directory
-      createFiles(fs, src, TOTAL_FILES);
-
-      // Track the number of copy operations
-      AtomicInteger copyCall = new AtomicInteger(0);
-      renameCrashInBetween(client, copyCall);
-      Assertions.assertThat(fs.rename(src, dst))
-          .describedAs("Rename should crash in between.")
-          .isFalse();
-
-      // Validate copy operation count
-      Assertions.assertThat(copyCall.get())
-          .describedAs("Copy operation count should be less than 10.")
-          .isLessThan(TOTAL_FILES);
-
-      // This will create marker file in the destination
-      fs.exists(dst);
-
-      copyCall.set(0);
-      // List Status on src, this will internally do rename recovery
-      triggerRenameRecovery(fs, src);
-
-      Assertions.assertThat(copyCall.get())
-          .describedAs("Copy operation count should be greater than 0.")
-          .isGreaterThan(0);
-
-      // Validate final state of destination and source
-      validateRename(fs, src, dst, false, true, false);
-    }
-  }
-
-  @Test
-  public void testRenameWhenAlreadyRenameRecoveryJsonFilePresent() throws Exception {
-    try (AzureBlobFileSystem fs = Mockito.spy(this.getFileSystem(
-        createConfig("5", "3", "2")))) {
-      assumeBlobServiceType();
-      AbfsBlobClient client = (AbfsBlobClient) addSpyHooksOnClient(fs);
-      fs.getAbfsStore().setClient(client);
-      Path src = new Path("/hbase/A1/A2");
-      Path dst = new Path("/hbase/A1/A3");
-
-      // Create sample files in the source directory
-      createFiles(fs, src, TOTAL_FILES);
-
-      // Track the number of copy operations
-      AtomicInteger copyCall = new AtomicInteger(0);
-      renameCrashInBetween(client, copyCall);
-      Assertions.assertThat(fs.rename(src, dst))
-          .describedAs("Rename should crash in between.")
-          .isFalse();
-
-      // Validate copy operation count
-      Assertions.assertThat(copyCall.get())
-          .describedAs("Copy operation count should be less than 10.")
-          .isLessThan(TOTAL_FILES);
-
-      // Validate final state of destination and source
-      validateRename(fs, src, dst, true, true, true);
-
-      copyCall.set(0);
-      Assertions.assertThat(fs.rename(src, dst))
-          .describedAs("Rename should succeed.")
-          .isTrue();
-
-      Assertions.assertThat(copyCall.get())
-          .describedAs("Copy operation count should be greater than 0.")
-          .isGreaterThan(0);
-
-      // Validate final state of destination and source
-      validateRename(fs, src, dst, false, true, false);
-    }
-  }
-
-  /**
    * Tests renaming a directory when the destination parent directory does not exist.
    * The test verifies that the rename operation fails as expected.
    *
@@ -2882,7 +2154,7 @@ public class ITestAzureBlobFileSystemRename extends
     try (AzureBlobFileSystem fs = this.getFileSystem()) {
       Path src = new Path("/A1/A2");
       Path dst = new Path("/A3/A4");
-      createFiles(fs, src, 1);
+      fs.create(new Path(src, "file.txt"));
 
       Assertions.assertThat(fs.rename(src, dst))
           .describedAs("Rename should fail as destination parent not exist.")
@@ -2945,7 +2217,7 @@ public class ITestAzureBlobFileSystemRename extends
     try (AzureBlobFileSystem fs = this.getFileSystem()) {
       Path src = new Path("/A1/A2");
       Path dst = new Path("/");
-      createFiles(fs, src, TOTAL_FILES);
+      fs.create(new Path(src, "file.txt"));
 
       Assertions.assertThat(fs.rename(src, dst))
           .describedAs("Rename should succeed.")
@@ -2994,54 +2266,4 @@ public class ITestAzureBlobFileSystemRename extends
            }
          });
    }
-
-  /**
-   * Triggers rename recovery by calling getPathStatus on the source path.
-   * This simulates a scenario where the rename operation was interrupted,
-   * and the system needs to recover the state of the source path.
-   *
-   * @param fs The AzureBlobFileSystem instance.
-   * @param src The source path to trigger recovery on.
-   * @throws Exception If an error occurs during the recovery process.
-   */
-  private void triggerRenameRecovery(AzureBlobFileSystem fs, Path src) throws Exception {
-    // Trigger rename recovery
-    TracingContext tracingContext = new TracingContext(
-        getConfiguration().getClientCorrelationId(), fs.getFileSystemId(),
-        FSOperationType.GET_FILESTATUS, TracingHeaderFormat.ALL_ID_FORMAT, null);
-    AzureServiceErrorCode errorCode = LambdaTestUtils.intercept(
-        AbfsRestOperationException.class, () -> {
-          fs.getAbfsStore().getClient().getPathStatus(src.toUri().getPath(), true,
-          tracingContext, null);
-    }).getErrorCode();
-    Assertions.assertThat(errorCode)
-        .describedAs("Path had to be recovered from atomic rename operation.")
-        .isEqualTo(PATH_NOT_FOUND);
-  }
-
-  /**
-   * Simulates a failure during the rename operation by throwing an exception
-   * when the copyBlob method is called. This is used to test the behavior of
-   * the rename recovery operation when a blob already exists at the destination.
-   *
-   * @param client The AbfsBlobClient instance.
-   * @param copyCall The AtomicInteger to track the number of copy calls.
-   * @throws AzureBlobFileSystemException If an error occurs during the operation.
-   */
-  private void renameCrashInBetween(AbfsBlobClient client, AtomicInteger copyCall)
-      throws AzureBlobFileSystemException {
-    Mockito.doAnswer(copyRequest -> {
-      if (copyCall.get() == FAILED_CALL) {
-        throw new AbfsRestOperationException(
-            BLOB_ALREADY_EXISTS.getStatusCode(),
-            BLOB_ALREADY_EXISTS.getErrorCode(),
-            BLOB_ALREADY_EXISTS.getErrorMessage(),
-            new Exception());
-      }
-      copyCall.incrementAndGet();
-      return copyRequest.callRealMethod();
-    }).when(client).copyBlob(Mockito.any(Path.class),
-        Mockito.any(Path.class), Mockito.nullable(String.class),
-        Mockito.any(TracingContext.class));
-  }
 }
